+++ src/main/java/tcbv/zhaohui/moon/enums/PresaleStage.java
@@ -3,6 +3,8 @@ package tcbv.zhaohui.moon.enums;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 /**
  * @author: zhaohui
  * @Title: PresaleStage
@@ -12,15 +14,15 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum PresaleStage {
-    PRESALE_ROUND_1(1, 500, 0.01),
-    PRESALE_ROUND_2(2, 400, 0.015),
-    PRESALE_ROUND_3(3, 400, 0.02),
-    PRESALE_ROUND_4(4, 350, 0.025),
-    PRESALE_ROUND_5(5, 350, 0.03),;
+    PRESALE_ROUND_1(1, 500, BigDecimal.valueOf(0.01)),
+    PRESALE_ROUND_2(2, 400, BigDecimal.valueOf(0.015)),
+    PRESALE_ROUND_3(3, 400, BigDecimal.valueOf(0.02)),
+    PRESALE_ROUND_4(4, 350, BigDecimal.valueOf(0.025)),
+    PRESALE_ROUND_5(5, 350, BigDecimal.valueOf(0.03));
 
     private final int stage;
 
     private final int number;
 
-    private final double price;
+    private final BigDecimal price;
 }
